git: Add helper to delete a (squash-)merged feature branch
When squash merging a feature branch, one can't delete it with git
branch -d $branch, since git will complain that it didn't get merged
yet. One needs to either delete it with git branch -D or rebase it first
to origin/master (to make sure it got indeed squash-merged as is) and
then delete it.

This script allows me to just do git finish when a branch got merged. To
avoid accidentally deleting a branch, we do interactive rebase which
should normally just show a 'noop'.